Supply model/format directly to hook
Most of the time people want to trigger on content model, or content
format. This allows triggering correctly even if you are viewing an
old revision with a different content model, or are using a content
format that is different than the default for the current model.

Also remove back compat code for pre 1.23
Change-Id: I9706eaae0983e2b7b99f83c9564ea04c5ec252ac