Use core behaviour regarding saving options for non-registered users
mw.Api#saveOptions will skip the API call and return a 'notloggedin'
error when the current user is not registered.

To handle temporary users as well, just fall back to that behaviour
instead of being explicit, and the right thing will also happen
(probably it will return an error, although the behaviour may change).

Bug: T332430
Change-Id: I4aafd6f74b74971c52b6222f8ec97a4963bb6a7b