Hooks: add public method to check conflicts with CodeMirror
While CodeMirror is still in beta, we need to leave CodeEditor as a
fallback when BetaFeatures is enabled and the CodeMirror beta feature
preference is not set. See T373711#11018957 for more information.

Additionally check $wgCodeMirrorV6 directly for the situation where that
is set and BetaFeatures is not installed or the CM beta feature is not
on the allow list.

Also only return early if the mode in $wgCodeEditorEnabledModes is set
to false. This allows for other modes to be handled by CodeEditor,
without having to list them all in the configuration.

Bug: T373711
Follow-Up: I0e4146a7fe9808e2c2d142f28fcc94c42acb0e3d
Change-Id: I3fd84e6d333e35650cced5bbe782a3cb57284345